requirements.txt
qq-botpy>=1.1.4
python-dotenv>=1.0.0
loguru>=0.7.0
httpx>=0.25.0
